backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-speech==2.21.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.0
